% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_convert_1.R
\name{convert_afl_1}
\alias{convert_afl_1}
\title{Convert the 3D simulation moment data into diffuse and direct light intensity}
\usage{
convert_afl_1(
  treescene_file,
  globrad,
  lat,
  lon,
  sensor_size = 1,
  inclination = 0,
  rotation = 0
)
}
\arguments{
\item{treescene_file}{path to the file of the app output for single moment simulation of a scene with trees}

\item{globrad}{Incoming global radiation in W m-2}

\item{lat}{Latitude of the field in °}

\item{lon}{Longitude of the field in °}

\item{sensor_size}{surface area of a sensor in the sensor field in m²}

\item{inclination}{Inclination setting (Geography) of the field simulation in degrees}

\item{rotation}{Rotation setting (Geography) of the field simulation in degrees}
}
\value{
tibble with variables datetime, pos_x, pos_y, direct_rad, diffuse_rad, total_rad
}
\description{
Convert the 3D simulation moment data into diffuse and direct light intensity
}
