% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_convert_ts.R
\name{convert_afl_ts}
\alias{convert_afl_ts}
\title{Convert the 3D simulation time series data into diffuse, direct and total light intensity}
\usage{
convert_afl_ts(
  treescene_dir_file,
  treescene_diff_file,
  datetime,
  globrad,
  lat,
  lon,
  sensor_size = 1,
  inclination = 0,
  rotation = 0,
  out_components = F
)
}
\arguments{
\item{treescene_dir_file}{path to the file of the app output for direct light}

\item{treescene_diff_file}{path to the file of the app output for diffuse light}

\item{datetime}{Single or multiple DateTime. Can be a character in UTC (YYYY-MM-DD HH:mm:ss) or a POSIXct}

\item{globrad}{Incoming global radiation in W m-2 corresponding to \code{datetime}. This should have a length equal to \code{datetime}.}

\item{lat}{Latitude of the field in °}

\item{lon}{Longitude of the field in °}

\item{sensor_size}{surface area of a sensor in the sensor field in m²}

\item{inclination}{Inclination setting (Geography) of the field simulation in degrees}

\item{rotation}{Rotation setting (Geography) of the field simulation in degrees}

\item{out_components}{if the output should return direct and diffuse components of radiation (default = F, for speed)}
}
\value{
tibble with variables datetime, pos_x, pos_y, total_rad
}
\description{
Convert the 3D simulation time series data into diffuse, direct and total light intensity
}
\details{
The simulation output should be generated using the application available at \link{https://agroforestry.ugent.be}
for a full year. The output is restricted to the time period for which globrad data were provided.
\code{sensor_size} is the result of (Sensor_Size_x x Sensor_Size_y) / (Sensors_Count_x x Sensors_Count_y)
}
