% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_1.R
\name{convert_data_1}
\alias{convert_data_1}
\title{Convert the 3D simulation moment data into diffuse and direct light intensity}
\usage{
convert_data_1(conv_factors, treescene_file)
}
\arguments{
\item{conv_factors}{a tibble as produced by the \code{make_conv_factors_1} function}

\item{treescene_file}{path to the file of the app output for single moment simulation of a scene with trees}
}
\value{
tibble with variables datetime, pos_x, pos_y, direct_rad, diffuse_rad, total_rad
}
\description{
Convert the 3D simulation moment data into diffuse and direct light intensity
}
