% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ts.R
\name{convert_data_ts}
\alias{convert_data_ts}
\title{Convert the 3D simulation time series data into diffuse, direct and total light intensity}
\usage{
convert_data_ts(
  conv_factors,
  treescene_dir_file,
  treescene_diff_file,
  lat,
  lon
)
}
\arguments{
\item{conv_factors}{a tibble as produced by the \code{make_conv_factors_ts} function}

\item{treescene_dir_file}{path to the file of the app output for direct light}

\item{treescene_diff_file}{path to the file of the app output for diffuse light}

\item{lat}{Latitude of the field in °}

\item{lon}{Longitude of the field in °}
}
\value{
tibble with variables datetime, pos_x, pos_y, total_rad
}
\description{
Convert the 3D simulation time series data into diffuse, direct and total light intensity
}
\details{
The simulation output should be generated using the application available at \link{https://agroforestry.ugent.be}
for a full year. The output is restricted to the time period for which globrad data were provided for the \code{make_conv_factors_ts} function.
}
