% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solartotime.R
\name{convert_solar_utc}
\alias{convert_solar_utc}
\title{Date and time conversion for input in agroforestry app}
\usage{
convert_solar_utc(solar_day, solar_time, year = 2020, lat, lon)
}
\arguments{
\item{solar_day}{Day used as input in the app (integer between 0-365), where 0 is 21 December}

\item{solar_time}{Hour used as input in the app (0-23.99), where 12 corresponds to solar noon}

\item{year}{Year for which the output is needed}

\item{lat}{Latitude of the field in °}

\item{lon}{Longitude of the field in °}
}
\value{
\code{datetime}: in UTC
}
\description{
convert the time in UTC to solar time where 12:00 is solar noon and
convert the date to solar day where 0 = December 21
}
\examples{
convert_solar_utc(solar_day = 180, solar_time = 12, lat = 50.5, lon = 3.8)
}
