% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todiel.R
\name{day_to_time_rad}
\alias{day_to_time_rad}
\title{Approximate higher time resolution radiation from daily data}
\usage{
day_to_time_rad(date, glob_rad, lat, lon, timeres_out)
}
\arguments{
\item{date}{the dates for which you want the conversion in "YYYY-MM-DD" format. Should have the same length as \code{glob_rad}}

\item{glob_rad}{the daily radiation sum in MJ d-1. Should have the same length as \code{doy}}

\item{lat}{the latitude where the data are from}

\item{lon}{the longitude the data are from}

\item{timeres_out}{the time resolution in \code{h} requested for the output}
}
\value{
a tibble with variables \code{datetime} and \code{radiation}
}
\description{
Approximate higher time resolution radiation from daily data
}
