% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdir.R
\name{diffdir_fun}
\alias{diffdir_fun}
\title{Calculate the diffuse and direct component of solar radiation on a horizontal plane}
\usage{
diffdir_fun(datetime, globrad, lat, lon)
}
\arguments{
\item{datetime}{Single or multiple DateTime. Can be a character in UTC (YYYY-MM-DD HH:mm:ss) or a POSIXct}

\item{globrad}{Incoming global radiation in W m-2. Its length should match length of \code{datetime}.}

\item{lat}{Latitude of the field in °}

\item{lon}{Longitude of the field in °}
}
\value{
A tibble with:
\itemize{
\item datetime: DateTime object
\item globrad: global radiation in W m-2
\item frac_diff: fraction of diffuse light in global radiation (0-1)
\item rad_diff: diffuse component of global radiation in W m-2
\item rad_dir: direct component of global radiation in W m-2
\item theta: sun angle from the zenith in radians (0 at the zenith and PI/2 at the horizon)
\item phi: sun azimuth in radians (direction along the horizon, measured from north to east), e.g. 0 is north and $\frac{\pi}{4}$ northeast
}
}
\description{
Calculate the diffuse and direct component of solar radiation on a horizontal plane
}
\examples{
out <- diffdir_fun(globrad = 600, lat = 50.95, lon = 3.8, datetime = "2022-06-21 12:00:00")
}
