% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_app}
\alias{launch_app}
\title{Launch the agroforlight interactive app}
\usage{
launch_app()
}
\value{
launches the agroforlight app at \url{https://agroforestry.ilvo.be/agroforlight/} in your standard browser
}
\description{
Launch the agroforlight interactive app
}
\examples{
launch_app()

}
