% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeconv_1.R
\name{make_conv_factors_1}
\alias{make_conv_factors_1}
\title{Calculates conversion factors for a single simulation}
\usage{
make_conv_factors_1(emptyscene_file, datetime, globrad, lat, lon)
}
\arguments{
\item{emptyscene_file}{Path to the file of the app output for an empty scene}

\item{datetime}{Single DateTime. Can be a character in UTC (YYYY-MM-DD HH:mm:ss) or a POSIXct}

\item{globrad}{Incoming global radiation in W m-2}

\item{lat}{Latitude of the field in °}

\item{lon}{Longitude of the field in °}
}
\value{
a tibble with two rows for following variables:
\itemize{
\item \code{light}: "sunlight" and "diffuse light"
\item \code{refvalue}: reference value used for light conversion
\item \code{app_day}: Day used for the \url{https://agroforestry.ugent.be} app (0 = December 21)
\item \code{app_hour}: Time of day used for the \link{https://agroforestry.ugent.be} app (12 = solar noon)
\item \code{app_pheno}: Leaf growth stage for the \link{https://agroforestry.ugent.be} app (0 = no leaves, 1 = all leaves fully developed)
\item \code{radiation}: direct and diffuse radiation fractions( W m-2)
\item \code{convfactor}: conversion factor for direct light (sunlight) and diffuse light
}
}
\description{
Calculates conversion factors for a single simulation
}
\examples{
make_conv_factors_1(emptyscene_file = "moment_180_12_100_ref.csv", lat = 50.8, lon = 3.8, globrad = 500, datetime = "2024-03-21 12:00:00")
}
