% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeconv_ts.R
\name{make_conv_factors_ts}
\alias{make_conv_factors_ts}
\title{Calculates conversion factors for a time series (year) simulation}
\usage{
make_conv_factors_ts(
  emptyscene_dir_file,
  emptyscene_diff_file,
  datetime,
  globrad,
  lat,
  lon
)
}
\arguments{
\item{emptyscene_dir_file}{Path to the file of the app output for direct light}

\item{emptyscene_diff_file}{Path to the file of the app output for diffuse light}

\item{datetime}{Single or multiple DateTime. Can be a character in UTC (YYYY-MM-DD HH:mm:ss) or a POSIXct}

\item{globrad}{Incoming global radiation in W m-2}

\item{lat}{Latitude of the field in °}

\item{lon}{Longitude of the field in °}
}
\value{
a tibble with the variables datetime, ConvFactorDiffuse, ConvFactorDirect
}
\description{
Calculates conversion factors for a time series (year) simulation
}
\details{
The simulation output should be generated using the application available at \link{https://agroforestry.ugent.be}
for a full year starting at "\code{start_year}-12-21". The output is restricted to the time period for which \code{globrad} data are provided.
}
