% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{make_phenology_file}
\alias{make_phenology_file}
\title{Write a phenology file for input to the 3-D simulation application}
\usage{
make_phenology_file(
  filename,
  doy_start_flush,
  doy_stop_flush,
  doy_start_fall,
  doy_stop_fall,
  shape = "linear"
)
}
\arguments{
\item{filename}{Path to the filename where the phenology file will be written}

\item{doy_start_flush}{Day of the year of onset of leaf appearance}

\item{doy_stop_flush}{Day of the year when all leaves have fully developed}

\item{doy_start_fall}{Day of the year when leaves start to fall}

\item{doy_stop_fall}{Day of the year when all leaves have fallen}

\item{shape}{shape of the phenology development. See \emph{details}}
}
\value{
writes a csv file
}
\description{
Write a phenology file for input to the 3-D simulation application
}
\details{
The day of the year inputs are transformed into the proper day for input in the application available at \link{https://agroforestry.ugent.be},
where the 'solar year' starts on December 21. Therefore, in the output, the onset of leaf development will appear 11 days later.

For now only a linear shape is available.
}
\examples{
make_phenology_file("pheno_01.csv", 120, 134, 280, 310)
}
