% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetosolar.R
\name{timetosolar}
\alias{timetosolar}
\title{Date and time conversion for input in agroforestry app}
\usage{
timetosolar(datetime, lat, lon)
}
\arguments{
\item{datetime}{Single DateTime. Can be a character in UTC (YYYY-MM-DD HH:mm:ss) or a POSIXct}

\item{lat}{Latitude of the field in °}

\item{lon}{Longitude of the field in °}
}
\value{
a list with:
\itemize{
\item \code{day_for_app}: the day number to input in the \link{https://agroforestry.ugent.be} app
\item \code{time_for_app}: the Time of day to input in the \link{https://agroforestry.ugent.be} app
}
}
\description{
convert the time in UTC to solar time where 12:00 is solar noon and
convert the date to solar day where 0 = December 21
}
\examples{
timetosolar("2024-06-21 12:00:00", lat = 50.5, lon = 3.8)
}
